refactor: 수동 로또 구매 0일 경우 구매 번호 입력 없음
